Don't crash if CanvasUniformLocation is 0
CanvasContext::getUniformLocation returns 0, if the
shader program does not declare the requested uniform
(as defined in the WebGL spec). That does mean, we have to
check for 0 in the uniform* functions to avoid crashes.

This is already done for most of the uniform* functions. This
commit introduces the remaining checks.

Change-Id: If123b5026e9a878913621c0bbba81bf77e5a433f
Reviewed-by: Tomi Korpipää <tomi.korpipaa@digia.com>
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>