Fix some ES versioning logic
Multisampling is fully available in ES 3.0+ so extend the condition.
(well technically we should check for the relevant extensions on GLES2
too since embedded/mobile devices will have it, but let's leave that for later)

Similarly, forcing ES 2.0 for the shared context when the original was ES 3.x is
not necessarily a good idea. So do like desktop GL and request the same version.

Change-Id: I71b3c8fac44a5cde8dee7fe8e77f451ff63c8589
Reviewed-by: Tomi Korpipää <tomi.korpipaa@digia.com>
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>