Fix context creation with threaded render loop
Creating a context that shares with a context that is current on another
thread is not safe, some drivers can reject this.

We work this around by astutely introducing an additional context that
lives on the same thread as the scenegraph's context, shares with it, but
is never current. The main thread's WebGL context will then share with
this extra context.

Task-number: QTBUG-44677
Change-Id: Ic3ba739c8cda4f4382705c19ed452e8e374c73d7
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>