Fix textureandlight resource usage
We have an alias for gl-matrix.js in the qrc so we have to match that.
Right now gl-matrix.js is not loaded and thus mat4 is not usable.

We have to have the absolute path since otherwise we are relative to
the main qml file which is in a qml subdir in case of this example.

Change-Id: Ie16f9e44a0a9d0498dd34270e2eb8df9efe4a92d
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>