Fixes issues in the examples.
Fixes the look'n'feel of the examples to be in line with the new Qt logo and website.
Instead of changing the simple texture and light example it was decided to add a new example
that shows how to combine QtQuick 2D UI with three.js based 3D rendering.

Change-Id: Ie3a21b03f8aa850ceae93252a764bd0d3264e229
Task-number: QTBUG-45696
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>