Execute Canvas3D OpenGL in render thread
The render loop has been redesigned.
The OpenGL commands are queued instead of executed immediately.
updatePaintNode() now transfers the commands to the renderer,
which implements all OpenGL related things that used to be in
the Canvas class. The actual rendering is done in response to
beforeRendering signal in the render thread.

[ChangeLog][Renderer] OpenGL rendering moved to render thread.

Task-number: QTBUG-44953
Change-Id: Ibd5897e64b84b6c7ce169cece8293381eca94ab9
Reviewed-by: Tomi Korpipää <tomi.korpipaa@theqtcompany.com>
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>