Fix QML type naming to avoid clashes with Qt 3D and other modules
There should be no API impact for changing the names of uncreatable
types. ContextAttributes3D is creatable, but most likely nobody
actually creates it, so it was deemed safe to change.
Some of the old type names differed from the documentation,
which is now also fixed as a side effect.

Change-Id: I3ec36d69e324fee22afa09a8db218f8e5beb93dc
Task-number: QTBUG-46105
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>