Actual fix for "shaking" bug that occurred in some GPUs
The first FBO we create gets corrupted by something outside
Canvas3D in some cases. E.g. changing the text in an completely unrelated
Text element may corrupt Canvas3D FBOs. This only happens with some GPU
drivers, likely due to a bug in the drivers themselves. To work around
the issue, the first FBO we create is a dummy that we delete once the
real FBOs are done.

Change-Id: Ie8c2a0ec3899c3eb22e4de8809c3b41607cc83c3
Task-number: QTBUG-46101
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>