Tweaking the shininess of the planets in Planets example.
The shininess of the planets in the example were left to the default 30.0 which is
quite shiny. To achieve more realistic look'n'feel the planets with specular texture
(earth) are tweaked to be a bit more shiny and the planets without a specual texture
(the rest) are tweaked to be not shiny at all.

Change-Id: I01acf6789cc632c8fba9be80a0792ae9170d05fd
Reviewed-by: Titta Heikkala <titta.heikkala@theqtcompany.com>
Reviewed-by: Pasi Keränen <pasi.keranen@digia.com>