Fix two GL errors and warnings related to resizing
Resizing window so that canvas height is zero shouldn't attempt
to recreate framebuffers, as it will fail. We can just draw to
the old framebuffers in those cases.
Another GL error happened when rapidly resizing the window,
as main thread directly adjusted the renderer FBO size.
Moved the adjustment to the sync() method.

Change-Id: Ib35c8439057d52e64186c7321208bb61ae94f58d
Reviewed-by: Tomi Korpipää <tomi.korpipaa@theqtcompany.com>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@theqtcompany.com>