Align pixel-row of the image passed to glTexImage2D to UNPACK_ALIGN
From what I can understand from the OpenGL specification, the image array
that is passed as the last parameter to glTexImage2D, the pixel width must
always be align to a boundary of what UNPACK_ALIGN parameter is set to.
For example if UNPACK_ALIGN is set to 4 (default) the width of the
image is 501 pixels, each pixel 3 bytes, each pixel row in image the
array must take 1504 bytes.

Task-number: QTBUG-54876
Change-Id: I7e3ea0bb11422903a6ca30eeb3ab930b9588f22f
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Pasi Keränen <pasi.keranen@qt.io>