Eradicate Q_FOREACH loops and mark the module as Q_FOREACH-free
In CanvasContext::setContextLostState(), ported from iterating
over keys() to iterating over the container itself. This seems
to be pretty safe, as it doesn't look like there's recursing
into the CanvasContext happening.

In CanvasGlCommandQueue::getCanvasId(), replaced a convoluted
loop with QMap::key(value, default) (required implementing the
equality relational operators for GlResource).

In ~StaticFactoryMapDeleter(), replaced the Q_FOREACH with a
call to qDeleteAll().

Change-Id: I9b849cce6e8ff48b746c682895b00e0d5982bacb
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>