CanvasTextureImageFactory: fix quadratic loop in notifyLoadedImages()
The old code uses a Java-style iterator on a QList, but
failed to use the iterator's remove() function, opting
instead to call QList::removeOne(). Both would have been
quadratic, but the current code is even worse in that it
re-scans the list for the element 'it' currently points
at.

Fix by using std::remove_if, which is linear, and, since
it doesn't take a copy of the container, also avoids a
deep copy of 'm_loadingImagesList' on actual removal (it
will attempt to detach unconditional now, due to the use
of non-const begin()/end(), but the variable is now no
longer shared with anyone, so there will actually be no
detach.

Change-Id: I71cd08bd40994e86a6470fdd9bcd55c18913b8d8
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>