Don't overwrite execute command queue when transferring queued commands
There's no guarantee that we'll only transfer commands after
successfully executing the existing command in the queue, so we
must take care to append to the execute queue, not overwrite it.

This can for example happen if an update request comes in which
results in transferring commands via Canvas::sync(), followed by
a synchronous command such as readPixels, which will trigger
another transfer plus an execution of the command queue.

Task-number: QTBUG-62963
Change-Id: Id0f326e98477ca14e70ad05706c62a8bf8a8d5f2
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>