Doc: Resolve all documentation warnings
QDoc in Qt 5.11 uses Clang to parse C++ documentation - even though
Qt Canvas doesn't expose a C++ API, we need to pass a module header
to Clang, as well as disable the custom module-specific namespace
for documentation builds. This is because we still parse the .cpp
files and get documentation warnings for classes that contain
/*! .. */ documentation comments, unless we explicitly mark the
classes \internal.

Change-Id: Ia7ee26b55d232c724bec78f0c3bf4018b5102913
Reviewed-by: Martin Smith <martin.smith@qt.io>